Add Subcollections::Settings
This code has been living in the base controller, when it really should
just exist as a subcollection, so moving it there so it is matching
conventions (easier to find) and to better represent what it does.
